Add weather icon to display and refactor
Added displayWeatherIcon() which displays a weather icon depending on
the current weather condition.

DisplayErrorMsg() and hideErrorMsg() now change the visibility of the
error message instead of its inner text.

DisplayData() now changes the visibility of the div with a class of
“main” and the toggle units button.

displayTemps() and displayWinds() are now exported for later use in
addListeners.js